requirements.txt
Pillow>=10.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
Jinja2>=3.1.0
requests>=2.31.0
click>=8.1.0
pytest>=7.4.0
